+++ pinokio.js
@@ -63,7 +63,7 @@ module.exports = {
         { icon: "fa-solid fa-microphone-lines", text: "Outputs", href: "app/output?fs=true" },
         { icon: "fa-solid fa-folder-tree", text: "HF Cache", href: "cache/HF_HOME/hub?fs=true" },
         { icon: "fa-regular fa-circle-xmark", text: "Uninstall Startup Service", href: "unservice.js" },
-        { icon: "fa-solid fa-rotate", text: "Update & Restart", href: "update_and_restart.js" }
+        { icon: "fa-solid fa-rotate", text: "Update", href: "update.js" }
       ]
     }
 
@@ -95,7 +95,7 @@ module.exports = {
             href: "open_external.js",
             params: { url: browserUrl } },
           { icon: "fa-solid fa-terminal", text: "Terminal", href: "start.js" },
-          { icon: "fa-solid fa-rotate", text: "Update & Restart", href: "update_and_restart.js" },
+          { icon: "fa-solid fa-rotate", text: "Update", href: "update.js" },
           { icon: "fa-solid fa-folder-tree", text: "HF Cache", href: "cache/HF_HOME/hub?fs=true" },
           { icon: "fa-solid fa-microphone-lines", text: "Outputs", href: "app/output?fs=true" },
           { icon: "fa-solid fa-wand-magic-sparkles",